Improve zoom performance
Waveform now is no longer resampled and drawn to canvas at all when
zooming. Only on mouse up. All zooming when mouse is down dragging on
zoom slider is achieved by scaling the x axis with css transform. All
waveform zoom levels are pre calculated when the waveform is loaded, and
they are stored in the ZoomedWaveform component. Opacity and scale is
then applied on each one when required.